fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12187 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/windows/GUIDialog_GLObjChooser.cpp
@@ -194,7 +194,7 @@ GUIDialog_GLObjChooser::onCmdFilter(FXObject*, FXSelector, void*) {
         }
     }
     myList->clearItems();
-    const int numSelected = selectedGlIDs.size();
+    const int numSelected = (const int)selectedGlIDs.size();
     for (int i = 0; i < numSelected; i++) {
         myList->appendItem(selectedMicrosimIDs[i], flag, (void*) & (*myIDs.find(selectedGlIDs[i])));
     }
